Move subscription into namespace
