--- app/models/subscription.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-# Models a subscription of Gather product itself.
-class Subscription < ApplicationRecord
-  acts_as_tenant :cluster
-
-  belongs_to :community, inverse_of: :subscription
-
-  def empty?
-    stripe_id.nil? && quantity.nil?
-  end
-
-  def registered?
-    stripe_id.present?
-  end
-end
