+++ app/views/layouts/_footer.html.erb
@@ -10,9 +10,9 @@
         &nbsp;
         <%= link_to("Settings", admin_settings_url(type: "community")) %>
       <% end %>
-      <% if policy(Subscription.new(community: current_community)).show? %>
+      <% if policy(Subscription::Subscription.new(community: current_community)).show? %>
         &nbsp;
-        <%= link_to("Subscription", subscription_url) %>
+        <%= link_to("Subscription", subscription_subscription_url) %>
       <% end %>
       <% if policy(Billing::Account.new(community: current_community)).index? %>
         &nbsp;
