+++ spec/models/subscription/subscription_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-describe Subscription do
+describe Subscription::Subscription do
   it "has a valid factory" do
     create(:subscription)
   end
